Left main container test working
Rendered left main container and then got the
iterate button with getby text. Then we reset
the the onclick function to the handle mock
iterate function. Key was changing to onclick
lower case and not camel case since we already
rendered the left main container component
so the iterate button was an html element.
Big typescript win

Co-author-by: Mike <mdtagg91@gmail.com>
Co-author-by: Aiden <aidencarere@gmail.com>
Co-author-by: Mel <mkoppens12@hotmail.com>
Co-author-by: Dylan <dster777@gmail.com>